Fix sly now takes own Ticket count into account for evaluation

+++ hpp/Entity.hpp
@@ -60,7 +60,7 @@ namespace tlk
 
             for (const Connection& con : options)
             {
-                uint score = scoreCon(con);
+                double score = scoreCon(con);
                 if (bestScore < score)
                 {
                     best = &con;
